Create Testing Segment 1
This is used for quick calculations for the corner coordinates of Segment 1. This can be verified through the SolidWorks model in the shared J Drive: Finishing Old Paper -> 3D Curve ->Segment 1. Just change the parameters in the code and make sure they match in the text file used for equations in SolidWorks (Finishing Old Paper -> Parameters). Then use the measure function in SolidWorks, relative to Coordinate System2.